save gas

+++ contracts/string_utils.sol
@@ -699,7 +699,7 @@ contract StringUtils{
         uint addressUint = uint(uint160(x));
         for (uint i = 0; i < 20; i++) {
             uint8 b = uint8(addressUint / (2**(152 - i*8))); // 152 == 8*(20-1)
-            uint8 index = uint8(2*i);
+            uint index = 2*i;
             s[index] = char(b / 16);
             s[index+1] = char(b % 16);            
         }
